fix #2
